KEYLOGGER
Software-Based Keyloggers:
These are programs or scripts that are installed on a computer or device to capture keystrokes. They can be written in various programming languages, including Python. Software-based keyloggers can work in the background, hidden from the user's view, making them difficult to detect. They can capture keystrokes and then save the recorded data to a file, send it to a remote server, or perform other actions as programmed,Due to their potential for misuse and unethical behavior, the creation and use of software-based keyloggers for malicious purposes is generally considered illegal and unethical. It's important to respect user privacy and follow legal regulations.