Create automessagebot
***AutoMessageBot***
AutoMessageBot is a Python script designed to automate sending repetitive messages. It utilizes the pyautogui library to type a specified message and press 'Enter' a specified number of times. This script is intended to be run in the command prompt (cmd) for optimal functionality.

Usage
Open your command prompt (cmd).

Navigate to the directory containing the script.

Run the script using the following command:

bash
Copy code
python automessagebot.py
Enter the message you want to send.

Input the number of times you want to send the message.

The script will automatically set the cursor on the message tab, type the message, and press 'Enter' the specified number of times.

Disclaimer: Please use this script responsibly and only in compliance with the terms of service of the application you are using it with.

Requirements
Python 3.x
pyautogui library